+++ Models/DTO/Addproduct.cs
@@ -28,7 +28,7 @@ namespace Loza.Models.DTO
         [StringLength(24)]
         public string Color { get; set; }
 
-        public int ColorNo { get; set; }
+        public long ColorNo { get; set; }
         [Required]
         [Range(0, 100)]
         public decimal Quantity { get; set; } = 0;
